refactor(adapters): relocate platform-generic modules out of adapters/discord (#1786)
Moves attachment-types.ts, attachments.ts, and channel-context.ts out of
src/adapters/discord/ into src/runner/ and src/common/ respectively — they
consume only platform-neutral types (AttachmentInfo, InboundMessage config)
and have no discord.js dependency, so core's dispatch-handler no longer
needs to import from a platform adapter directory.

Behavior-preserving move: git mv + import-path rewrites only, no logic
changes. Tests relocated alongside their modules.
